bitvavo-dashboard/requirements.txt
ccxt>=4.0.0
pandas>=2.1
numpy>=1.26
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.24
psycopg[binary]>=3.1
